Split server/db into per-domain modules
Break server/db/index.ts (1008 lines) into domain files, keeping
index.ts as the barrel that re-exports them so "../db" imports stay
stable:

- client (createDb factory, chunkForD1)
- scans (scan lifecycle, decisions, risk-summary helpers)
- organizations, npm-connections, rate-limit, events

Collapse the overlapping risk-summary readers into one shared helper
in scans.ts.

Pure refactor, no behavior change.
